use s{s}ize_t to handle large files

+++ src/cache/path_status_cache.h
@@ -4,6 +4,7 @@
 #include <mutex>
 #include <string>
 #include <thallium.hpp>
+#include <stddef.h>
 
 namespace tl = thallium;
 
@@ -16,7 +17,7 @@ using PathStatusMDCacheTableVal = std::optional<int>;
 class PathStatusCache final : public TLCache<PathStatusCacheTableKey, PathStatusMDCacheTableVal> {
     public:
     bool check_and_put_force(const Key& key);
-    void update_cache_status(const Key&, int status);
+    void update_cache_status(const Key&, ssize_t status);
     int wait_on_cache_status(const Key& key);
 
     tl::condition_variable cv;
